+++ odoo/local-src/sf_purchase/models/purchase.py
@@ -11,6 +11,10 @@ class Purchase(models.Model):
         'res.partner',
         string='Secondary Vendor',
         help='An indirect supplier.')
+    confirmation_date = fields.Datetime(
+        string='Confirmation Date',
+        readonly=True, index=True,
+        help="Date on which the purchase order is confirmed.")
     print_draft = fields.Boolean(
         string='Print Draft',
         help='Print draft stamp on the top of the request for quotation '
@@ -19,7 +23,11 @@ class Purchase(models.Model):
     @api.multi
     def button_confirm(self):
         res = super(Purchase, self).button_confirm()
-        self.print_draft = False
+        self.write({
+            'confirmation_date': fields.Datetime.now(),
+            'print_draft': False
+        })
+
         return res
 
     @api.multi