+++ odoo/local-src/sf_rma/models/mrp_repair.py
@@ -2,15 +2,22 @@
 # Copyright 2017 Camptocamp SA
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl)
 
-from odoo import fields, models
+from odoo import fields, models, api
 
 
 class MRPRepair(models.Model):
     _inherit = 'mrp.repair'
 
+    @api.depends('state')
+    def _compute_user_id(self):
+        for repair in self:
+            repair.user_id = self.env.uid
+
     rma_id = fields.Many2one('sf.rma', string="RMA")
 
-    user_id = fields.Many2one('res.users', string="Technician")
+    user_id = fields.Many2one('res.users', string="Technician",
+                              compute='_compute_user_id', store=True,
+                              track_visibility='onchange')
     planned_date = fields.Date()
     note_resolution = fields.Text(string="Resolution note")
     image_arrival = fields.Binary(attachment=True)