+++ odoo/local-src/sf_rma_sale_order/models/res_company.py
@@ -9,4 +9,13 @@ class ResCompany(models.Model):
 
     _inherit = 'res.company'
 
-    rma_service_product_id = fields.Many2one('product.product')
+    rma_service_service_product_id = fields.Many2one(
+        'product.product')
+
+    rma_service_consumable_product_id = fields.Many2one(
+        'product.product')
+
+    rma_service_stockable_product_id = fields.Many2one(
+        'product.product')
+
+    rma_service_additional_description = fields.Char()