+++ odoo/songs/install/data_demo.py
@@ -44,7 +44,10 @@ def import_countries(ctx):
 @anthem.log
 def import_crm_team(ctx):
     """ Importing sales team / channel from csv """
-    load_csv(ctx, 'data/demo/crm_team.csv', 'crm.team')
+    model = ctx.env['crm.team'].with_context({
+        'tracking_disable': True,
+    })
+    load_csv(ctx, 'data/demo/crm_team.csv', model)
 
 
 @anthem.log
@@ -56,13 +59,19 @@ def import_location(ctx):
 @anthem.log
 def import_customers(ctx):
     """ Importing customers from csv """
-    load_csv(ctx, 'data/demo/customers.csv', 'res.partner')
+    model = ctx.env['res.partner'].with_context({
+        'tracking_disable': True,
+    })
+    load_csv(ctx, 'data/demo/customers.csv', model)
 
 
 @anthem.log
 def import_suppliers(ctx):
     """ Importing suppliers from csv """
-    load_csv(ctx, 'data/demo/suppliers.csv', 'res.partner')
+    model = ctx.env['res.partner'].with_context({
+        'tracking_disable': True,
+    })
+    load_csv(ctx, 'data/demo/suppliers.csv', model)
 
 
 @anthem.log
@@ -74,7 +83,10 @@ def import_drone_type(ctx):
 @anthem.log
 def import_product(ctx):
     """ Importing products from csv """
-    load_csv(ctx, 'data/demo/product.csv', 'product.template')
+    model = ctx.env['product.template'].with_context({
+        'tracking_disable': True,
+    })
+    load_csv(ctx, 'data/demo/product.csv', model)
 
 
 @anthem.log
@@ -129,6 +141,38 @@ def import_purchase_order(ctx):
 
 
 @anthem.log
+def import_waves(ctx):
+    """ Importing waves from csv """
+    load_csv(ctx, 'data/demo/wave.csv', 'stock.picking.wave')
+
+
+@anthem.log
+def import_pricelist(ctx):
+    """ Importing pricelists from csv """
+    model = ctx.env['product.pricelist'].with_context({
+        'tracking_disable': True,
+    })
+    load_csv(ctx, 'data/demo/pricelist.csv', model)
+    model_item = ctx.env['product.pricelist.item'].with_context({
+        'tracking_disable': True,
+    })
+    load_csv(ctx, 'data/demo/pricelist_item.csv', model_item)
+
+
+@anthem.log
+def import_invoices_supplier(ctx):
+    """ Importing invoices supplier from csv """
+    model = ctx.env['account.invoice'].with_context({
+        'tracking_disable': True,
+    })
+    load_csv(ctx, 'data/demo/invoice_supp_head.csv', model)
+    model_item = ctx.env['account.invoice.line'].with_context({
+        'tracking_disable': True,
+    })
+    load_csv(ctx, 'data/demo/invoice_supp_line.csv', model_item)
+
+
+@anthem.log
 def main(ctx):
     """ Loading demo data """
     import_users(ctx)
@@ -138,10 +182,12 @@ def main(ctx):
     import_customers(ctx)
     import_suppliers(ctx)
     import_crm_team(ctx)
+    import_waves(ctx)
     import_location(ctx)
     import_customers(ctx)
     import_drone_type(ctx)
     import_product(ctx)
+    import_pricelist(ctx)
     import_serial_number(ctx)
     import_workcenter(ctx)
     import_rma_cause(ctx)
@@ -149,4 +195,5 @@ def main(ctx):
     import_bank_account(ctx)
     import_sales_order(ctx)
     import_purchase_order(ctx)
+    import_invoices_supplier(ctx)
     return