+++ odoo/songs/install/data_full.py
@@ -76,21 +76,28 @@ def import_countries(ctx):
 @anthem.log
 def import_customers(ctx):
     """ Importing customers from csv """
-    load_csv(ctx, 's3://prod-sf-odoo-data/install/customers.csv',
-             'res.partner')
+    model = ctx.env['res.partner'].with_context({
+        'tracking_disable': True,
+    })
+    load_csv(ctx, 's3://prod-sf-odoo-data/install/customers.csv', model)
 
 
 @anthem.log
 def import_suppliers(ctx):
     """ Importing suppliers from csv """
-    load_csv(ctx, 's3://prod-sf-odoo-data/install/suppliers.csv',
-             'res.partner')
+    model = ctx.env['res.partner'].with_context({
+        'tracking_disable': True,
+    })
+    load_csv(ctx, 's3://prod-sf-odoo-data/install/suppliers.csv', model)
 
 
 @anthem.log
 def import_crm_team(ctx):
     """ Importing sales team / channel from csv """
-    load_csv(ctx, 's3://prod-sf-odoo-data/install/crm_team.csv', 'crm.team')
+    model = ctx.env['crm.team'].with_context({
+        'tracking_disable': True,
+    })
+    load_csv(ctx, 's3://prod-sf-odoo-data/install/crm_team.csv', model)
 
 
 @anthem.log
@@ -110,8 +117,10 @@ def import_drone_type(ctx):
 @anthem.log
 def import_product(ctx):
     """ Importing products type from csv """
-    load_csv(ctx, 's3://prod-sf-odoo-data/install/product.csv',
-             'product.template')
+    model = ctx.env['product.template'].with_context({
+        'tracking_disable': True,
+    })
+    load_csv(ctx, 's3://prod-sf-odoo-data/install/product.csv', model)
 
 
 @anthem.log
@@ -171,6 +180,51 @@ def import_purchase_order(ctx):
 
 
 @anthem.log
+def import_bom(ctx):
+    """ Importing bill of materials from csv """
+    load_csv(ctx, 's3://prod-sf-odoo-data/install/mrp_bom.csv', 'mrp.bom')
+    load_csv(ctx, 's3://prod-sf-odoo-data/install/mrp_bom_line.csv',
+             'mrp.bom.line')
+
+
+@anthem.log
+def import_waves(ctx):
+    """ Importing waves from csv """
+    load_csv(ctx, 's3://prod-sf-odoo-data/install/wave.csv',
+             'stock.picking.wave')
+
+
+@anthem.log
+def import_pricelist(ctx):
+    """ Importing pricelists from csv """
+    model = ctx.env['product.pricelist'].with_context({
+        'tracking_disable': True,
+    })
+    load_csv(ctx, 's3://prod-sf-odoo-data/install/pricelist.csv',
+             model)
+    model_item = ctx.env['product.pricelist.item'].with_context({
+        'tracking_disable': True,
+    })
+    load_csv(ctx, 's3://prod-sf-odoo-data/install/pricelist_item.csv',
+             model_item)
+
+
+@anthem.log
+def import_invoices_supplier(ctx):
+    """ Importing invoices supplier from csv """
+    model = ctx.env['account.invoice'].with_context({
+        'tracking_disable': True,
+    })
+    load_csv(ctx, 's3://prod-sf-odoo-data/install/invoice_supp_head.csv',
+             model)
+    model_item = ctx.env['account.invoice.line'].with_context({
+        'tracking_disable': True,
+    })
+    load_csv(ctx, 's3://prod-sf-odoo-data/install/invoice_supp_line.csv',
+             model_item)
+
+
+@anthem.log
 def main(ctx):
     """ Loading full data """
     import_users(ctx)
@@ -184,7 +238,9 @@ def main(ctx):
     import_crm_team(ctx)
     import_location(ctx)
     import_drone_type(ctx)
+    import_waves(ctx)
     import_product(ctx)
+    import_pricelist(ctx)
     import_serial_number(ctx)
     import_workcenter(ctx)
     import_rma_cause(ctx)
@@ -192,4 +248,6 @@ def main(ctx):
     import_bank_account(ctx)
     import_sales_order(ctx)
     import_purchase_order(ctx)
+    import_bom(ctx)
+    import_invoices_supplier(ctx)
     return